backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
PyMySQL==1.1.0
PyJWT==2.8.0
bcrypt==4.1.2
